Harden clock runtime storage save path and tidy manifest/page details
Agent-Logs-Url: https://github.com/Victor-Muthomi/taska/sessions/9d070a56-c1ad-40a7-8316-c30dcedbb17f

+++ lib/features/clock/presentation/pages/clock_services_page.dart
@@ -394,8 +394,7 @@ class _ClockServicesPageState extends ConsumerState<ClockServicesPage>
     restoredAlarms.sort((a, b) => a.nextRingAt.compareTo(b.nextRingAt));
 
     final restoredTimerDuration = Duration(
-      seconds: state.timerDurationSeconds.clamp(0, _maxTimerDurationSeconds)
-          .toInt(),
+      seconds: state.timerDurationSeconds.clamp(0, _maxTimerDurationSeconds),
     );
     final restoredEndsAtUtc = state.timerEndsAtUtcIso == null
         ? null
